web: Add a dark theme to the web player
A dark theme has been added to the web player. Whether dark or light
mode is used is automatically recognised and determines which theme is
used.
This changes the appearance of the context menu, the modals, the panic
screen and the message overlay.

Additionally, the CSS code has been simplified and some unnecessary
parts have been removed.